Handle Office Open XML types for filed generated outside MSOffice
It turns out that the position of '[Content_Types].xml' is not fixed in
this format, and whereas MS Office adds it in the beginning in general,
when files are generated using Google Docs, or RubyXL, this is not
necessarily true. They might add it at the end, so all these would be
detected as application/zip.

This fixes #39